requirements.txt
meson==1.6.1
packaging==24.2
wheel @ file:///opt/homebrew/Cellar/python%403.13/3.13.1/libexec/wheel-0.45.1-py3-none-any.whl#sha256=da46333d5dcbde6e20cf7e2f8fff9e9ce76e8c94dc4afd6fb95fc4bc2745fb5e
fastapi==0.104.1
uvicorn==0.24.0
firecrawl-py==0.0.16
openai==1.3.7
resend==0.6.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
apscheduler==3.10.4
pytz==2023.3
